+++ pilot/pkg/networking/core/v1alpha3/gateway_test.go
@@ -36,12 +36,12 @@ import (
 	"istio.io/istio/pilot/pkg/networking/util"
 	"istio.io/istio/pilot/pkg/security/model"
 	memregistry "istio.io/istio/pilot/pkg/serviceregistry/memory"
+	"istio.io/istio/pilot/test/xdstest"
 	"istio.io/istio/pkg/config/constants"
 	"istio.io/istio/pkg/config/mesh"
 	"istio.io/istio/pkg/config/schema/collections"
 	"istio.io/istio/pkg/config/schema/gvk"
 	"istio.io/istio/pkg/proto"
-	"istio.io/istio/pkg/test/xdstest"
 )
 
 func TestBuildGatewayListenerTlsContext(t *testing.T) {
