+++ pilot/test/xdstest/extract.go
